LiteE_Sense8_system: v2.0-ALPHA
Modified the script, added upstream for 3.16.709.15 (SUPER LAGGY), fix the HTC Sense launcher crashing and remove the ability to set it as default, making Lawnchair the default launcher, added ScreenCapture (for screenshots, not recording videos, scroll feature doesn't seem to work tho), also modified the bootanimation but needs to be reworked.

+++ build_liteealpha.sh
@@ -1,6 +1,7 @@
 # /bin/bash
-
+# Version 2.0-ALPHA
 LITEE_PATH="/home/j6idot/HTC/LiteE10"
+UPSTREAM_VERSION=3.16.709.15
 
 clear
 
@@ -9,28 +10,46 @@ sudo umount -l *
 # end
 
 # Create a backup of the system image
-read -p "Would you like to create a backup of the system image? (system.img) (y/n) " RESP
-if [ "$RESP" = "y" ]; then
+if [ ! -f "system_backup.img" ]; then
+	read -p "Would you like to create a backup of the system image? (system.img) (y/n) " RESP
+	if [ "$RESP" = "y" ]; then
 	echo - Creating a backup as system_backup.img
 	cp system.img system_backup.img
+	fi
+else 
+	sleep 3
+fi
+# end
+
+# Mount the system partition
+echo -- Mounting the system partition
+mkdir system &> /dev/zero
+e2fsck -yf system.img > $LITEE_PATH/e2fsck.log
+resize2fs system.img 4000M > $LITEE_PATH/resize2fs.log
+sudo mount -o loop system.img system || { echo 'mount failed!' ; exit 1; }
+df -h system/
+echo "Before:" > $LITEE_PATH/system_size.log
+df -h system/ >> $LITEE_PATH/system_size.log
+sleep 1
+# end
+
+# Upstream to specified version
+echo --- Upstream to $UPSTREAM_VERSION
+if [ ! -d "$LITEE_PATH/LiteE/upstream/$UPSTREAM_VERSION" ]; then
+  echo "The specified upstream folder does not exist. Are you sure you inserted the right version?"
+  exit
+fi
+read -p "Would you like to upstream the ROM to $UPSTREAM_VERSION? (y/n) " RESP
+if [ "$RESP" = "y" ]; then
+		sudo cp -r $LITEE_PATH/LiteE/upstream/$UPSTREAM_VERSION/system/* $LITEE_PATH/system/
+		sudo cp $LITEE_PATH/LiteE/upstream/$UPSTREAM_VERSION/system/build.prop $LITEE_PATH/system/
 else 
 	sleep 1
 fi
 
 # end
 
-# Mount the system partition
-echo -- Mounting the system partition
-mkdir system
-e2fsck -yf system.img > /dev/zero
-resize2fs system.img 4000M
-sudo mount -o loop system.img system || { echo 'mount failed!' ; exit 1; }
-df -h system/
-sleep 1
-# end
-
 # Debloat /system/app
-#	MyHTC \
 echo --- Debloating /system/app
 cd system/app
 sudo rm -rf AndroidPay \
@@ -38,6 +57,7 @@ sudo rm -rf AndroidPay \
 	Duo \
 	EditorsDocs \
 	facebook-stub \
+	FaceLock \
 	Gmail2 \
 	Hangouts \
 	HTCSpeakCyberon \
@@ -63,7 +83,6 @@ cd .. && cd priv-app
 sudo rm -rf AMapNetworkLocation \
 	AndroidHtcSync \
 	AppCloud \
-	EasyAccessService \
 	Facebook \
 	Frisbee \
 	GooglePartnerSetup \
@@ -83,7 +102,21 @@ sudo rm -rf AMapNetworkLocation \
 	Velvet \
 	DemoFLOPackageInstaller \
 	FlexNet \
-	installer
+	installer \
+	HtcLockScreen \
+	HtcWeatherClockWidget \
+	HtcIceView \
+	HtcBIDHandler \
+	HtcAccessoryService \
+	UIBC \
+	ContextualWidget \
+	CtsShimPrivPrebuilt \
+	CustomizationSettingsProvider \
+	CustomizationSetup \
+	Default_IME_Provider \
+	DemoFLOPackageInstaller \
+	DrawingBoard \
+	HTCAdvantage
 cd ..
 # end
 
@@ -95,9 +128,17 @@ sudo chmod 0644 $LITEE_PATH/system/app/LatinIMEGooglePrebuilt/LatinIMEGooglePreb
 sudo chown $ROOT:$ROOT $LITEE_PATH/system/app/LatinIMEGooglePrebuilt/LatinIMEGooglePrebuilt.apk
 # end
 
+# copy ScreenCapture to /system/app
+echo --- Copying ScreenCapture to /system/app
+sudo cp -r $LITEE_PATH/LiteE/system/app/ScreenCapture $LITEE_PATH/system/app
+sudo chmod 0755 $LITEE_PATH/system/app/ScreenCapture
+sudo chmod 0644 $LITEE_PATH/system/app/ScreenCapture/ScreenCapture.apk
+sudo chown $ROOT:$ROOT $LITEE_PATH/system/app/ScreenCapture/ScreenCapture.apk
+# end
+
 # install lawnchair
 echo --- Install Lawnchair
-#sudo cp $LITEE_PATH/LiteE/system/priv-app/Prism/Prism.apk $LITEE_PATH/system/priv-app/Prism
+sudo cp $LITEE_PATH/LiteE/system/priv-app/Prism/Prism.apk $LITEE_PATH/system/priv-app/Prism
 sudo mkdir $LITEE_PATH/system/priv-app/Lawnchair
 sudo cp -r $LITEE_PATH/LiteE/system/priv-app/Lawnchair $LITEE_PATH/system/priv-app/
 sudo chmod -R 0755 $LITEE_PATH/system/priv-app/Lawnchair/
@@ -125,15 +166,19 @@ sudo rm -rf qmi-framework-tests \
 	diag_socket_log \
 	diag_uart_log \
 	init.usbdiag.sh \
+	ipacm-diag \
 	mmi_debug \
 	mmi_diag \
 	mm-qjpeg-dec-test \
 	mm-qjpeg-enc-test \
 	mm-qomx-idec-test \
 	mm-vidc-omx-test \
+	mm-qomx-ienc-test \
+	PktRspTest \
 	qjpeg-dma-test \
 	qmi_simple_ril_test \
-	ssr_diag
+	ssr_diag \
+	test_diag
 
 sudo rm -rf $LITEE_PATH/system/vendor/bin/init.minidump.sh
 sudo rm -rf $LITEE_PATH/system/bin/hrdump_v2
@@ -144,12 +189,14 @@ sudo rm -rf $LITEE_PATH/system/bin/net_log
 sudo rm -rf $LITEE_PATH/system/bin/qsee_logger
 sudo rm -rf $LITEE_PATH/system/bin/opendiag
 sudo rm -rf $LITEE_PATH/system/bin/mlogger
+sudo rm -rf $LITEE_PATH/system/bin/fmfactorytest
+sudo rm -rf $LITEE_PATH/system/bin/fmfactorytestserver
 sudo cp -r $LITEE_PATH/LiteE/system/vendor/etc/init/hw* $LITEE_PATH/system/vendor/etc/init/hw
 # end
 
 # change the splash screen
 echo --- Edit the 2nd splash screen
-sudo cp $LITEE_PATH/LiteE/system/customize/resource/* $LITEE_PATH/system/customize/resource/
+sudo cp $LITEE_PATH/LiteE/system/customize/resource/splash_screen.png $LITEE_PATH/system/customize/resource/
 # end
 
 # disable signature verification
@@ -212,18 +259,27 @@ sudo rm -rf $LITEE_PATH/system/customize/resource/quicklaunch_*
 # unmount the system and resize the image
 echo ---- Resizing the image
 cd $LITEE_PATH
-df -h system
+echo >> $LITEE_PATH/system_size.log
+echo "After:" >> $LITEE_PATH/system_size.log
+df -h system/ >> $LITEE_PATH/system_size.log
 sudo umount system
-e2fsck -yf system.img > /dev/zero
-resize2fs system.img 2900M
+echo >> $LITEE_PATH/e2fsck.log
+echo >> $LITEE_PATH/resize2fs.log
+e2fsck -yf system.img >> $LITEE_PATH/e2fsck.log
+resize2fs system.img 2700M >> $LITEE_PATH/resize2fs.log
+echo 
+echo 
+echo -----
 echo Done!
 echo Remember:
-echo chcon -R u:object_r:system_file:s0 /system/media/audio/notification
+echo chcon -R u:object_r:system_file:s0 /system/app/LatinIMEGooglePrebuilt/
+echo chcon -R u:object_r:system_file:s0 /system/app/ScreenCapture/
+echo chcon u:object_r:system_file:s0 /system/etc/permissions/com.google.android.dialer.support.xml
+echo chcon u:object_r:system_file:s0 /system/framework/com.google.android.dialer.support.jar
+echo chcon -R u:object_r:system_file:s0 /system/media/audio/notifications
 echo chcon -R u:object_r:system_file:s0 /system/media/audio/ringtones
 echo chcon -R u:object_r:system_file:s0 /system/media/audio/alarms
 echo chcon -R u:object_r:system_file:s0 /system/media/audio/ui
-echo chcon -R u:object_r:system_file:s0 /system/app/LatinIMEGooglePrebuilt/
 echo chcon -R u:object_r:system_file:s0 /system/priv-app/Lawnchair/
 echo chcon -R u:object_r:system_file:s0 /system/priv-app/Dialer/
-echo chcon u:object_r:system_file:s0 /system/framework/com.google.android.dialer.support.jar
-echo chcon u:object_r:system_file:s0 /system/etc/permissions/com.google.android.dialer.support.xml
+echo chcon u:object_r:system_file:s0 /system/vendor/etc/permissions/android.hardware.vulkan.compute-0.xml
